demo/requirements.txt
Django==2.2
django-rest-knox==4.0.0
djangorestframework==3.9.2
